Get rid of console.log statements.
Calling JSON.stringify on random objects is potentially dangerous for the
health of the application. In particular, if the object has a string whose
size is many megabytes we might run out of memory before even calling
console.log.

I got rid of the console.log calls entirely because they're not free and they
fill up the console for everybody, even if they're not developing the app.
Getting rid of JSON.stringify(note) was the important change.