Don't roundtrip through JSON to normalize objects.
I had to guess a little bit as to what "normalizeObj" was intended to do but
this seems to work (i.e. I can still sync notes) and it's much, much cheaper.
The object that was causing problems had a string that was many megabytes in
length (28 to be exact) and we were running out of memory deserializing it.

As far as I can tell, the only thing the roundtrip does is strip out undefined
values and functions from the final object, so this accomplishes that by on
serializing the properties that we actually want.